+++ src/app/UI/Accessory/Admin/Form/Controls/EditorJs/EditorJsTypeEnum.php
@@ -15,5 +15,6 @@ enum EditorJsTypeEnum:string
     case Audio = 'audio';
     case Gallery = 'gallery';
     case MultiImage = 'multiImage';
+    case TextStyle = 'textStyle';
 
 }